backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
starlette==0.27.0
anthropic==0.8.1
chromadb==0.4.18
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
gitpython==3.1.40
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
httpx==0.25.2
requests==2.31.0
numpy==1.26.2
typing-extensions==4.8.0
pytest==7.4.3
black==23.12.0
packaging==23.2
tqdm==4.66.1